+++ src/main/kotlin/racingcar/domain/RacingCarDriveResult.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 class RacingCarDriveResult(
     val carNumber: Long,
