+++ src/main/kotlin/racingcar/domain/RacingCarGame.kt
@@ -1,4 +1,8 @@
-package racingcar
+package racingcar.domain
+
+import racingcar.Car
+import racingcar.view.RacingCarGameInputView
+import racingcar.view.RacingCarGameResultView
 
 class RacingCarGame {
