+++ src/main/kotlin/racingcar/domain/RacingCarGameWinnerFinder.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 object RacingCarGameWinnerFinder {
