+++ src/main/kotlin/racingcar/domain/RacingCarRoundResult.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 class RacingCarRoundResult(
     val round: Long,
