+++ src/main/kotlin/racingcar/domain/RacingGameCarFactory.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 object RacingGameCarFactory {
