+++ src/main/kotlin/racingcar/view/RacingCarGameInputView.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.view
 
 class RacingCarGameInputView {
