+++ src/test/kotlin/racingcar/RacingGameCarFactoryTest.kt
@@ -4,6 +4,8 @@ import org.assertj.core.api.Assertions
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
+import racingcar.domain.RacingGameCarFactory
+import racingcar.domain.RandomNumberGenerator
 
 class RacingGameCarFactoryTest {
