+++ src/test/kotlin/racingcar/RandomDriveCarTest.kt
@@ -4,6 +4,8 @@ import io.mockk.every
 import io.mockk.mockk
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.domain.RandomDriveCar
+import racingcar.domain.RandomNumberGenerator
 
 class RandomDriveCarTest {
