+++ src/test/kotlin/racingcar/RandomNumberGeneratorTest.kt
@@ -2,6 +2,7 @@ package racingcar
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
+import racingcar.domain.RandomNumberGenerator
 
 class RandomNumberGeneratorTest {
